Removed unnecessary tlc namespace
We don't have to share our test and name of this test case won't be conflicted.
